linux: drop ancient patches that never went upstream
The Ouya and PS3 remote patches were added over 10 years ago
with the promise to get them upstream but that never happened.

Drop them as they are constantly breaking on major kernel updates.

If the functionality may still be wanted see this as a wakeup call
to send the patches upstream and add backport patches once they
are accepted or add other ways to eg set remote repeat.

Signed-off-by: Matthias Reichl <hias@horus.com>