atf: fix build
This is a target package, and as such config/functions set all sorts of
"target compiler" options, but atf needs to be build with the
"kernel compiler", e.g. aarch64 and not armv8a.

So unset all the wrong envvars from the wrong toolchain to not confuse
the atf build system.