libxslt: update build options aligning with upstream
https://github.com/GNOME/libxslt/blob/master/NEWS

The non-standard EXSLT crypto extensions and support for dynamically
loaded plugins are now disabled by default. These features can be
enabled by passing --with-crypto or --with-plugins to configure.
In a future release, these features will be removed.

Debug output and the debugger are disabled by default and can be
enabled by passing --with-debug or --with-debugger.

was identified as unnecessary upon review of incorrect pkgconfig file
generation https://gitlab.gnome.org/GNOME/libxslt/-/issues/136