hyperhdr: use CMAKE_POLICY_VERSION_MINIMUM workaround to compile with cmake-4.0.0
This should be reverted when HyperHDR is upgraded from 21.0.0.0.
The fix was included in https://github.com/awawa-dev/HyperHDR/commit/6779fdf843a66a03d56485c23f70fca9b13fcceb

This was unable to be patched, as the github project has subprojects that would not
be included in the generated tar.gz file.