mesa: fix logic error in glvnd selection
Commit 9d2001f14ec7 ("Add Generic-gl device") introduced a subtle error
in the glvnd enablement logic that lead to the Generic-Legacy image
being built without glvnd which breaks nvidia support.

Unlike most programming languages "||" and "&&" have the same precedence
in shell and "A || B && C" get evaluated as "(A || B) && C" instead of
the intended "A || (B && C)".

Fix this by moving the "AND not-X11" check into the "GL" test and
adjust formatting to make the intention clear.

Signed-off-by: Matthias Reichl <hias@horus.com>