flex: use flex:host stage1flex when building scan.c from scan.l
The compile logic whilst building a CROSS compiled flex could be improved.
Given that we have already built flex:host and the necessary stage1flex,
just use it. Using the --disable-bootstrap logic in the make file which
would normally call sed, replace the sed call with the call to be previously
built stage1flex. Both the patch and the update to package.mk are required.